+++ client/src/components/Header/Header.css
@@ -1,0 +1,6 @@
+@import url('https://fonts.googleapis.com/css2?family=Lilita+One&display=swap');
+
+#title {
+    font-family: 'Lilita One', sans-serif;
+    font-size: 1.5em;
+}
