+++ client/src/components/index.js
@@ -1,1 +1,1 @@
-export {default as Header} from "./Header";
+export {default as Header} from "./Header/Header.jsx";
